🚚 (consent-api): Rename service dir to services
