+++ apps/consent-api/src/services/search.ts
